+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 Other:
   - Added a new example for electron with the remote API. Thanks [@SebastienWae](https://github.com/SebastienWae)
 
+Bugfixes:
+  - Updates to documents fail with GraphQL replication. [#1812](https://github.com/pubkey/rxdb/issues/1812). Thanks [@gautambt](https://github.com/gautambt)
+
 ### 8.7.4 (2 December 2019)
 
 Other: